performous-app-build.sh: fixes
Make the script more robust with regards to path of the script itself, as well as being able to specify a different path for the performous source and optionally, build without overwriting everything.

+++ osx-utils/performous-app-build.sh
@@ -3,15 +3,28 @@ set -o errexit
 # the very first step is to check that dylibbundler exists,
 # without it the bundle would be broken
 
+SOURCE="${BASH_SOURCE[0]}"
+while [ -h "${SOURCE}" ]
+	do # resolve $SOURCE until the file is no longer a symlink
+		DIR="$( cd -P "$( dirname "${SOURCE}" )" && pwd )"
+		SOURCE="$(readlink "${SOURCE}")"
+		[[ "${SOURCE}" != /* ]] && SOURCE="${DIR}/${SOURCE}" # if $SOURCE was a relative symlink, we need to resolve it relative to the path where the symlink file was located
+	done
+CURRDIR="$( cd -P "$( dirname "$SOURCE" )" && pwd )"
+
 test -z ${PREFIXDIR} && PREFIXDIR="/opt/local" # By default, the default prefix for macports, change this if you're using a different path or package manager.
 test -z ${DEPLOYMENT_TARGET} && DEPLOYMENT_TARGET="10.14" # Change this if you want to target a different version of macOS.
+test -z ${PERFORMOUS_SOURCE} && PERFORMOUS_SOURCE="${CURRDIR}/.." # Change this if using another copy of the source or if this file is not under path/to/performous/osx-utils/
 MAKE_JOBS=$(sysctl -n hw.ncpu)
 test -z ${CC} && CCPATH="/usr/bin/clang" # Path to system Clang, change if you want another compiler.
 test -z ${CXX} && CXXPATH="/usr/bin/clang++" # Path to system Clang, change if you want another compiler.
 
+args="$@"
+debug="^(--debug|-d)$"
+cleanbuild="^(--no-clean|-k)$"
 
-args=("$@")
-debug="^(--debug|--DEBUG|-d|-D)$"
+DEBUG=0
+PERFORMOUS_CLEAN_BUILD=1
 
 function asksure {
 	echo -n "$1"
@@ -39,13 +52,21 @@ function exists {
 	fi
 }
 
-function arewedebug {
-	if [[ ${args[0]} =~ ${debug} ]]
+function check_arguments {
+	shopt -s nocasematch
+	for arg in ${args}
+	do
+		if [[ ${arg} =~ ${debug} ]]
 		then
-			DEBUG=1
-	else
-		DEBUG=0
-	fi
+				DEBUG=1
+		else
+			if [[ ${arg} =~ ${cleanbuild} ]]
+			then
+				PERFORMOUS_CLEAN_BUILD=0
+			fi
+		fi
+	done
+	shopt -u nocasematch
 }
 
 function bundlelibs {
@@ -64,20 +85,21 @@ function bundlelibs {
 function createdmg {
 	if [ "${FANCY_DMG}" == 0 ]
 		then
-			ln -sf /Applications "${CURRDIR}/out/Applications"
-			rm "${CURRDIR}/out/.DS_Store"
+			ln -sf /Applications "${PERFORMOUS_SOURCE}/out/Applications"
+			rm "${PERFORMOUS_SOURCE}/osx-utils/out/.DS_Store"
 			/usr/bin/hdiutil create -ov -srcfolder out -volname Performous -fs HFS+ -fsargs "-c c=64,a=16,e=16" -format UDRW RWPerformous.dmg
 			/usr/bin/hdiutil convert -ov RWPerformous.dmg -format UDZO -imagekey zlib-level=9 -o Performous.dmg
 			rm -f RWPerformous.dmg
 			cd ..
 	elif [ "${FANCY_DMG}" == 1 ]
 		then
-			rm -f "${CURRDIR}/Performous.dmg"
-			appdmg "${CURRDIR}/resources/dmg_spec.json" "${CURRDIR}/Performous.dmg"
+			rm -f "${PERFORMOUS_SOURCE}/osx-utils/Performous.dmg"
+			appdmg "${PERFORMOUS_SOURCE}/osx-utils/resources/dmg_spec.json" "${PERFORMOUS_SOURCE}/osx-utils/Performous.dmg"
 	fi
 }
 
 function main {
+	printf "\n"
 	if [[ "${DEBUG}" = 1 ]]
 		then
 			echo "Will create bundle for debugging..."
@@ -101,7 +123,7 @@ function main {
 			else
 				if exists npm
 					then
-						if asksure "appdmg is not installed, would you like to install it? (y/n)"
+						if asksure "* appdmg is not installed, would you like to install it? (y/n)"
 							then
 								sudo npm install -g https://github.com/LinusU/node-appdmg.git && FANCY_DMG=1
 							else
@@ -116,19 +138,10 @@ function main {
 		fi
 	fi
 
-	SOURCE="${BASH_SOURCE[0]}"
-	while [ -h "${SOURCE}" ]
-		do # resolve $SOURCE until the file is no longer a symlink
-			DIR="$( cd -P "$( dirname "${SOURCE}" )" && pwd )"
-			SOURCE="$(readlink "${SOURCE}")"
-			[[ "${SOURCE}" != /* ]] && SOURCE="${DIR}/${SOURCE}" # if $SOURCE was a relative symlink, we need to resolve it relative to the path where the symlink file was located
-		done
-	CURRDIR="$( cd -P "$( dirname "$SOURCE" )" && pwd )"
-
 # first compile performous, build dir shouldn't exist at this stage
 
 # define Bundle structure
-	TEMPDIR="${CURRDIR}/out/Performous.app/Contents"
+	TEMPDIR="${PERFORMOUS_SOURCE}/osx-utils/out/Performous.app/Contents"
 	RESDIR="${TEMPDIR}/Resources"
 	LIBDIR="${RESDIR}/lib"
 	LOCALEDIR="${RESDIR}/Locales"
@@ -136,16 +149,28 @@ function main {
 	BINDIR="${TEMPDIR}/MacOS"
 	ETCDIR="${RESDIR}/etc"
 
-	rm -rf "${TEMPDIR}"
-	mkdir -p "${TEMPDIR}"
-
-	rm -rf ./build
-	mkdir build
-	cd build
-
+	printf "\n"
+	echo "--- Performous source: ${PERFORMOUS_SOURCE}"
+	echo "--- Build folder: ${PERFORMOUS_SOURCE}/build"
+	echo "--- Output folder: ${PERFORMOUS_SOURCE}/osx-utils/out"
+	printf "\n"
+	if [[ "${PERFORMOUS_CLEAN_BUILD}" == 1 ]]
+		then
+			echo "Wiping build and output bundle folders..."
+			printf "\n"
+			rm -rf "${PERFORMOUS_SOURCE}/build"
+			rm -rf "${TEMPDIR}"
+			mkdir -p "${TEMPDIR}"
+		else
+			echo "No-clean mode enabled; preserving previous build directory."
+			printf "\n"
+			rm -rf "${TEMPDIR}"
+			mkdir -p "${TEMPDIR}"
+	fi
+	
 	cmake \
 	  -DCMAKE_INSTALL_PREFIX="$TEMPDIR" \
-	  -DCMAKE_SYSTEM_PREFIX_PATH="${PREFIXDIR}" \
+	  -DCMAKE_PREFIX_PATH="${PREFIXDIR}" \
 	  -DCMAKE_BUILD_TYPE="${RELTYPE}" \
 	  -DENABLE_WEBSERVER="ON" \
 	  -DCMAKE_VERBOSE_MAKEFILE="ON" \
@@ -156,21 +181,23 @@ function main {
 	  -DSHARE_INSTALL="Resources" \
 	  -DLOCALE_DIR="Resources/Locales" \
 	  -DCMAKE_CXX_FLAGS="-Wall -Wextra" \
-	  -DCMAKE_OSX_ARCHITECTURES="x86_64" ../..
+	  -DCMAKE_OSX_ARCHITECTURES="x86_64" \
+	  -B "${PERFORMOUS_SOURCE}/build" \
+	  -S "${PERFORMOUS_SOURCE}"
 	
-	make -j${MAKE_JOBS} install # You can change the -j value in order to spawn more build threads.
+	make -C "${PERFORMOUS_SOURCE}/build" -j${MAKE_JOBS} install # You can change the -j value in order to spawn more build threads.
 
 # then create the rest of the app bundle
 
 	mv "${TEMPDIR}/bin" "${BINDIR}"
 
-	cp ../resources/performous-launcher "${BINDIR}"
+	cp "${PERFORMOUS_SOURCE}/osx-utils/resources/performous-launcher" "${BINDIR}"
 	if [[ "${DEBUG}" = 1 ]]
 		then
 			sed -i '' -e 's|"\${CURRDIR}\/performous"|"\${CURRDIR}\/performous" --log debug|g' "$BINDIR/performous-launcher" # enable debug logging.
 	fi
-	cp ../resources/performous.icns "${RESDIR}"
-	cp ../resources/Info.plist "${TEMPDIR}"
+	cp "${PERFORMOUS_SOURCE}/osx-utils/resources/performous.icns" "${RESDIR}"
+	cp "${PERFORMOUS_SOURCE}/osx-utils/resources/Info.plist" "${TEMPDIR}"
 
 	mkdir -p "${FRAMEWORKDIR}"
 	mkdir -p "${LIBDIR}"
@@ -203,6 +230,6 @@ function main {
 	fi
 }
 
-arewedebug
+check_arguments
 
 main
